Improve performance of CalculateCoverage.
Removed code parsing an unchanged value on every invocation of a DoFn.  The purpose of that code was to skip the creation of annotations for data beyond the boundaries specified by --references.  But --references controls what data is pulled in and can have multiple entries for the same contig, so the code was incorrect. The code now creates annotations for all data that it receives.

Change-Id: Icd2920c18759b5491bafb7c1aec0e112b5d70d95